Improves Endpoint Robustness
This patch focuses on improving link reliability.
There should be pretty much no way to lose the
TCP side of the downlink once activated now. It
also improves the UI by forcing the user to launch the
app by plugging in the device, getting rid of the
annoying security requests by the android
subsystem.