Implement full wakelock
Partial wakelock really wasn't fixing the slowdown
issue. This patch implements an easily modifiable,
settable wakelock and sets it to full screen wakelock.

To prevent the downlink from being disabled by a screen
tap in flight, the app requires a long-press on the stop button
to enact a stop now.